Link to XLib and XInput dynamically.
