+++ keyboard/build.gradle.kts
@@ -14,7 +14,7 @@ fun KotlinMultiplatformExtension.configureLinux() {
 
         main.cinterops.create("device") { defFile("src/linuxX64Main/cinterop/device.def") }
         main.cinterops.create("x11") { defFile("src/linuxX64Main/cinterop/x11.def") }
-        mavenPublication { artifactId = "${project.name}-kt-linuxx64" }
+        mavenPublication { artifactId = "${project.name}-kt-linuxX64" }
     }
 
     val linuxX64Main by sourceSets.getting
@@ -24,7 +24,7 @@ fun KotlinMultiplatformExtension.configureLinux() {
 }
 
 fun KotlinMultiplatformExtension.configureMingw() {
-    mingwX64 { mavenPublication { artifactId = "${project.name}-kt-mingwx64" } }
+    mingwX64 { mavenPublication { artifactId = "${project.name}-kt-mingwX64" } }
 
     val mingwX64Main by sourceSets.getting
     val mingwX64Test by sourceSets.getting { dependsOn(mingwX64Main) }
@@ -33,15 +33,8 @@ fun KotlinMultiplatformExtension.configureMingw() {
 }
 
 kotlin {
-    if (ideaActive) {
-        when {
-            HostManager.hostIsLinux -> configureLinux()
-            HostManager.hostIsMingw -> configureMingw()
-        }
-    } else {
-        configureLinux()
-        configureMingw()
-    }
+    configureLinux()
+    configureMingw()
 
     sourceSets {
         val commonMain by getting {
